+++ src/widget/box_widget.rs
@@ -16,7 +16,6 @@
 use graphics::*;
 use opengl_graphics::GlGraphics;
 use piston::input::*;
-use graphics::rectangle::Border;
 
 use crate::core::callbacks::*;
 use crate::core::point::{Point, Size};
@@ -58,7 +57,7 @@ impl BoxWidget {
             c.transform.clone(),
         );
 
-        Rectangle::new_border(color, (border / 2.0)).draw(
+        Rectangle::new_border(color, border / 2.0).draw(
             [
                 0.0 as f64 + (border / 2.0) as f64,
                 0.0 as f64 + (border / 2.0) as f64,