+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 - Changed demo so that it does not resize, as resize is very slow right now
 - Added Linux to build rules in TravisCI.
 - Fixed BoxWidget line drawing width being too wide (#194)
+- Fixed layout code for Horizontal Layout (#175)
 
 ## 0.4.6